Fix opencl
Addresses several issues with the opencl tests:
- Model object does not correctly identify whether model was compiled
with stan_opencl. Therefore, models were not recompiled within the tests
when recompilation is needed. Resolved this by supplying a different
exe_file for each tests so recompilation will always happen.
- Tests referenced a metadata attribute that was not defined.
I've added checks for opencl related attributes that are set when
opencl is functioning properly.
- Version test used a data input that was never created.